% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFPUArea.R
\name{calcFPUArea}
\alias{calcFPUArea}
\title{calc FPU area based on LUH2 datase}
\usage{
calcFPUArea(x)
}
\arguments{
\item{x}{An output from readStates function of this package}
}
\value{
list. Contains both a dataframe and a SpatVector with FPU level info.
}
\description{
calc FPU area based on LUH2 datase
}
\examples{
\dontrun{
calcFPUArea(readStates())
}
}
\author{
Abhijeet Mishra
}
