% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStates.R
\name{readStates}
\alias{readStates}
\title{readStates}
\usage{
readStates()
}
\value{
Cell area for LUHv2 land use types on 0.25 degree grid
}
\description{
readStates
}
\examples{
\dontrun{
readStates()
}
}
\author{
Abhijeet Mishra
}
